Tested GetDPS, addedequippableItems enum in Character child classes for equip logic

+++ Characters/Rogue.cs
@@ -32,7 +32,7 @@ namespace RPG_Characters
             throw new NotImplementedException();
         }
 
-        public override float GetDPS()
+        public override double GetDPS()
         {
             throw new NotImplementedException();
         }
